DEV-7: Initial refactoring to distinct services
